requirements.txt
art
requests
google-cloud-storage
colorama
glob2
tqdm
